Refactor and continued framing
I have refactored the code a little bit. I will be reinstalling the
functionality removed in a later commit. I have also added some
inheritance structures to allow specialization of the NI for either the
client or the server. It was easier to make things click if this
structure was in place.